chore(ui_storage): update example
