+++ packages/firebase_ui_storage/example/lib/src/apps.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:firebase_ui_storage_example/main.dart';
+import 'package:firebase_ui_storage_example/storybook.dart';
 import 'package:firebase_ui_storage_example/src/storage_image_app.dart';
 import 'package:flutter/material.dart';
